backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
djoser==2.2.0
django-filter==23.5
pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
reportlab==4.0.7
django-colorfield==0.11.0
drf-extra-fields==3.7.0
